% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartrank.R
\name{smartrank}
\alias{smartrank}
\title{Rank a vector based on either alphabetical or frequency order}
\usage{
smartrank(x, sort_by = c("alphabetical", "frequency"), desc = FALSE)
}
\arguments{
\item{x}{A numeric, character, or factor vector}

\item{sort_by}{Sort ranking either by "alphabetical" or "frequency" . Default is "alphabetical"}

\item{desc}{A logical indicating whether the ranking should be in descending ( TRUE ) or ascending ( FALSE ) order.
When input is numeric, ranking is always based on numeric order.}
}
\value{
The ranked vector
}
\description{
This function acts as a drop-in replacement for the base \code{rank()} function with the added option to:
\enumerate{
\item Rank categorical factors based on frequency instead of alphabetically
\item Rank in descending or ascending order
}
}
\examples{
smartrank(c("apple","banana","banana","orange","apple"))
smartrank(c("apple","banana","banana","orange","apple"),sort_by = "frequency")
}
